*Sucess* Compiles, links, and starts. Ready for run environment for regression testing.

+++ SDK1/IfSearch/CharacteristicColor.cpp
@@ -207,7 +207,7 @@ qreal CharacteristicColor::hslDistance(const QColor & qc,
                   qreal lWeight,
                   qreal sThreshold) const
 {
-    qreal h1, l1, s1, h2, l2, s2, result=0.0;
+    float h1, l1, s1, h2, l2, s2, result=0.0;
     if (qIsNull(hWeight + sWeight + lWeight))
         return -1.0;
     QColor(toRgb()).toHsl().getHslF(&h1, &s1, &l1);
